+++ src/player/main/sp/it/pl/layout/Factory.kt
@@ -131,6 +131,7 @@ class DeserializingFactory(val launcher: File): ComponentFactory<Component> {
 }
 
 class NodeFactory<T: Node>(val id: UUID, val type: KClass<out T>, override val name: String, val constructor: () -> T): ComponentFactory<Component>, Locatable {
+   val info: WidgetInfo? = type.companionObjectInstance?.asIf<WidgetInfo>()
    override suspend fun create() = APP.widgetManager.factories.getFactory("Node").orNone().create().apply { fieldsRaw["node"] = PropVal.PropVal1(type.jvmName) }
    override fun toString() = "${javaClass.simpleName} $name $type"
    override val location = APP.widgetManager.factories.getFactory("Node").orNone().location