+++ src/player/main/sp/it/pl/ui/objects/picker/IconPicker.kt
@@ -1,12 +1,13 @@
 package sp.it.pl.ui.objects.picker
 
+import de.jensd.fx.glyphs.GlyphIcons
 import javafx.geometry.Insets
 import javafx.geometry.Pos.CENTER
 import javafx.geometry.Side.RIGHT
 import javafx.scene.layout.Priority.ALWAYS
-import javafx.scene.text.Font
 import sp.it.pl.main.IconMA
 import sp.it.pl.main.IconMD
+import sp.it.pl.ui.nodeinfo.IconPickerContent
 import sp.it.pl.ui.objects.icon.Icon
 import sp.it.pl.ui.objects.window.popup.PopWindow
 import sp.it.util.access.v
@@ -16,8 +17,8 @@ import sp.it.util.ui.hBox
 import sp.it.util.ui.lay
 import sp.it.util.ui.vBox
 
-class FontPicker(onOk: (Font) -> Unit) {
-   val pickerContent = FontPickerContent()
+class IconPicker(onOk: (GlyphIcons?) -> Unit) {
+   val pickerContent = IconPickerContent()
    val popup = PopWindow()
    val editable = v(true)
 
@@ -26,11 +27,11 @@ class FontPicker(onOk: (Font) -> Unit) {
          padding = Insets(15.0)
          lay(ALWAYS) += pickerContent
          lay += hBox(15.0, CENTER) {
-            lay += Icon(IconMD.CHECK, 22.0).onClickDo { popup.hide(); if (editable.value) onOk(pickerContent.font) }.net {
+            lay += Icon(IconMD.CHECK, 22.0).onClickDo { popup.hide(); if (editable.value) onOk(pickerContent.selection.value) }.net {
                editable sync { isDisable = !it }
                it.withText(RIGHT, "Use")
             }
-            lay += Icon(IconMD.CHECK_ALL, 22.0).onClickDo { if (editable.value) onOk(pickerContent.font) }.net {
+            lay += Icon(IconMD.CHECK_ALL, 22.0).onClickDo { if (editable.value) onOk(pickerContent.selection.value) }.net {
                editable sync { isDisable = !it }
                it.withText(RIGHT, "Apply")
             }